refactor: modularise code based on review comments
