+++ src/data/repositories/SurveyFormD2Repository.ts
@@ -1,83 +1,41 @@
 import { D2Api } from "@eyeseetea/d2-api/2.36";
 import { D2TrackerEvent } from "@eyeseetea/d2-api/api/trackerEvents";
 import { Future } from "../../domain/entities/generic/Future";
-import {
-    Questionnaire,
-    QuestionnaireEntity,
-    QuestionnaireStage,
-} from "../../domain/entities/Questionnaire/Questionnaire";
-import { Id, Ref } from "../../domain/entities/Ref";
+import { Questionnaire } from "../../domain/entities/Questionnaire/Questionnaire";
+import { Id } from "../../domain/entities/Ref";
 import { SurveyRepository } from "../../domain/repositories/SurveyRepository";
 import { apiToFuture, FutureData } from "../api-futures";
 import _ from "../../domain/entities/generic/Collection";
 import { ImportStrategy, ProgramCountMap } from "../../domain/entities/Program";
-import { Survey, SURVEY_FORM_TYPES, SURVEY_STATUSES } from "../../domain/entities/Survey";
-import { DataValue } from "@eyeseetea/d2-api";
+import { Survey, SURVEY_FORM_TYPES } from "../../domain/entities/Survey";
 import {
     D2TrackerTrackedEntity,
     D2TrackerTrackedEntity as TrackedEntity,
     TrackedEntitiesGetResponse,
 } from "@eyeseetea/d2-api/api/trackerTrackedEntities";
 import {
-    D2TrackerEnrollment,
-    D2TrackerEnrollmentAttribute,
-} from "@eyeseetea/d2-api/api/trackerEnrollments";
-import {
     getChildProgramId,
     getParentDataElementForProgram,
-    getSurveyNameBySurveyFormType,
     getSurveyType,
     getTrackedEntityAttributeType,
     isTrackerProgram,
 } from "../utils/surveyProgramHelper";
 import {
-    programsWithRepeatableSections,
-    SURVEY_ID_DATAELEMENT_ID,
-    SURVEY_ID_PATIENT_DATAELEMENT_ID,
-    WARD_ID_DATAELEMENT_ID,
-    WARD2_ID_DATAELEMENT_ID,
-    AMR_SURVEYS_PREVALENCE_DEA_SURVEY_ID,
-    hiddenFields,
-    SURVEY_ID_FACILITY_LEVEL_DATAELEMENT_ID,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SSTF,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRL,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_PIS,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SRL,
-    AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRF,
     PREVALENCE_FACILITY_LEVEL_FORM_ID,
     PPS_WARD_REGISTER_ID,
     PPS_HOSPITAL_FORM_ID,
     PPS_PATIENT_REGISTER_ID,
-    keyToDataElementMap,
-    PPS_SURVEY_FORM_ID,
-    PREVALENCE_SURVEY_FORM_ID,
     SURVEY_NAME_DATAELEMENT_ID,
     PREVALENCE_SURVEY_NAME_DATAELEMENT_ID,
     PPS_COUNTRY_QUESTIONNAIRE_ID,
 } from "../entities/D2Survey";
+import { ProgramDataElement, ProgramMetadata } from "../entities/D2Program";
 import {
-    D2ProgramRule,
-    D2ProgramRuleAction,
-    D2ProgramRuleVariable,
-    Option,
-    Program,
-    ProgramDataElement,
-    ProgramMetadata,
-    ProgramStage,
-    ProgramStageSection,
-    TrackedEntityAttibute,
-} from "../entities/D2Program";
-import { QuestionnaireSection } from "../../domain/entities/Questionnaire/QuestionnaireSection";
-import { QuestionnaireRule } from "../../domain/entities/Questionnaire/QuestionnaireRules";
-import {
-    BooleanQuestion,
-    DateQuestion,
-    DateTimeQuestion,
-    NumberQuestion,
-    Question,
-    SelectQuestion,
-    TextQuestion,
-} from "../../domain/entities/Questionnaire/QuestionnaireQuestion";
+    mapProgramToQuestionnaire,
+    mapQuestionnaireToEvent,
+    mapQuestionnaireToTrackedEntities,
+} from "../utils/surveyFormMappers";
+import { mapEventToSurvey, mapTrackedEntityToSurvey } from "../utils/surveyListMappers";
 
 export class SurveyD2Repository implements SurveyRepository {
     constructor(private api: D2Api) {}
@@ -117,7 +75,7 @@ export class SurveyD2Repository implements SurveyRepository {
                             trackedEntity => {
                                 if (resp.programs[0] && trackedEntity) {
                                     return Future.success(
-                                        this.mapProgramToQuestionnaire(
+                                        mapProgramToQuestionnaire(
                                             resp.programs[0],
                                             undefined,
                                             trackedEntity,
@@ -143,7 +101,7 @@ export class SurveyD2Repository implements SurveyRepository {
                         return this.getEventProgramById(eventId).flatMap(event => {
                             if (resp.programs[0] && event) {
                                 return Future.success(
-                                    this.mapProgramToQuestionnaire(
+                                    mapProgramToQuestionnaire(
                                         resp.programs[0],
                                         event,
                                         undefined,
@@ -167,7 +125,7 @@ export class SurveyD2Repository implements SurveyRepository {
                 //return empty form
                 else {
                     return Future.success(
-                        this.mapProgramToQuestionnaire(
+                        mapProgramToQuestionnaire(
                             resp.programs[0],
                             undefined,
                             undefined,
@@ -189,447 +147,6 @@ export class SurveyD2Repository implements SurveyRepository {
         });
     }
 
-    private mapProgramToQuestionnaire(
-        program: Program,
-        event: D2TrackerEvent | undefined,
-        trackedEntity: TrackedEntity | undefined,
-        programDataElements: Ref[],
-        dataElements: ProgramDataElement[],
-        options: Option[],
-        programStages?: ProgramStage[],
-        programStageSections?: ProgramStageSection[],
-        trackedEntityAttributes?: TrackedEntityAttibute[],
-        programRules?: D2ProgramRule[],
-        programRuleVariables?: D2ProgramRuleVariable[],
-        programRuleActions?: D2ProgramRuleAction[]
-    ): Questionnaire {
-        //If the Program has sections, fetch and use programStageSections
-        const sections: QuestionnaireSection[] = programStageSections
-            ? programStageSections.map(section => {
-                  const { questions, sectionAddQuestion } = this.mapProgramDataElementToQuestions(
-                      isTrackerProgram(program.id),
-                      section.dataElements,
-                      dataElements,
-                      options,
-                      event,
-                      trackedEntity
-                  );
-
-                  return {
-                      title: section.name,
-                      code: section.id,
-                      questions: questions,
-                      isVisible: true,
-                      stageId: section.programStage.id,
-                      sortOrder: section.sortOrder,
-                      showAddQuestion: sectionAddQuestion,
-                  };
-              })
-            : //If the Program has no sections, create a single section
-              [
-                  {
-                      title: "Survey Info",
-                      code: "default",
-                      questions: this.mapProgramDataElementToQuestions(
-                          isTrackerProgram(program.id),
-                          programDataElements,
-                          dataElements,
-                          options,
-                          event,
-                          trackedEntity
-                      ).questions,
-                      isVisible: true,
-                      stageId: "default",
-                      sortOrder: 1,
-                  },
-              ];
-
-        //If the Program has stages, fetch and use programStages
-        const stages: QuestionnaireStage[] = programStages
-            ? programStages.map(stage => {
-                  const currentProgramStageSections =
-                      programStages.length === 1 //If there is only 1 program stage, then all the sections belong to it.
-                          ? sections
-                          : sections.filter(section => section.stageId === stage.id);
-
-                  //check if current program has repeatable Questions.
-                  if (programsWithRepeatableSections.find(p => p === program.id)) {
-                      const groupedProgramStageSections = _(currentProgramStageSections)
-                          .sortBy(s => s.sortOrder)
-                          .groupBy(section =>
-                              section.title.substring(0, section.title.lastIndexOf(" "))
-                          );
-
-                      const processedSections: QuestionnaireSection[] = [];
-                      groupedProgramStageSections.forEach(group => {
-                          if (group[1].length > 1) {
-                              const currentSectionGroup: QuestionnaireSection[] = group[1].map(
-                                  (section, index) => {
-                                      return {
-                                          ...section,
-                                          isVisible: index === 0 ? true : false,
-                                          showAddnew: true,
-                                      };
-                                  }
-                              );
-                              processedSections.push(...currentSectionGroup);
-                          } else processedSections.push(...group[1]);
-                      });
-
-                      return {
-                          title: stage.name,
-                          code: stage.id,
-                          sections: _(processedSections)
-                              .sortBy(section => section.sortOrder)
-                              .value(),
-                          isVisible: true,
-                          instanceId: trackedEntity?.enrollments
-                              ?.at(0)
-                              ?.events.find(e => e.programStage === stage.id)?.event,
-                      };
-                  } else {
-                      // no need for grouping and hiding logic
-                      return {
-                          title: stage.name,
-                          code: stage.id,
-                          sections: _(currentProgramStageSections)
-                              .sortBy(section => section.sortOrder)
-                              .value(),
-                          isVisible: true,
-                          instanceId: trackedEntity?.enrollments
-                              ?.at(0)
-                              ?.events.find(e => e.programStage === stage.id)?.event,
-                      };
-                  }
-              })
-            : //If the Program has no stages, create a single stage
-              [
-                  {
-                      title: "STAGE",
-                      code: "STAGE",
-                      sections: _(sections)
-                          .sortBy(section => section.sortOrder)
-                          .value(),
-                      isVisible: true,
-                  },
-              ];
-
-        const orgUnitId = isTrackerProgram(program.id)
-            ? trackedEntity?.orgUnit ?? ""
-            : event?.orgUnit ?? "";
-
-        const questionnaireRules: QuestionnaireRule[] = this.getProgramRules(
-            programRules,
-            programRuleVariables,
-            programRuleActions
-        );
-
-        const form: Questionnaire = {
-            id: program.id,
-            name: program.name,
-            description: program.name,
-            orgUnit: { id: orgUnitId },
-            year: "",
-            isCompleted: false,
-            isMandatory: false,
-            stages: stages.sort((a, b) => a.title.localeCompare(b.title, "en", { numeric: true })),
-            subLevelDetails: {
-                enrollmentId: trackedEntity
-                    ? trackedEntity.enrollments?.at(0)?.enrollment ?? ""
-                    : "",
-            },
-            rules: questionnaireRules,
-        };
-
-        if (trackedEntityAttributes) {
-            const profileQuestions: Question[] = this.mapTrackedAttributesToQuestions(
-                trackedEntityAttributes,
-                options,
-                trackedEntity
-            );
-
-            const profileSection: QuestionnaireEntity = {
-                title: "Profile",
-                code: "PROFILE",
-                questions: profileQuestions,
-                isVisible: true,
-                stageId: "PROFILE",
-            };
-            form.entity = profileSection;
-        }
-        return form;
-    }
-
-    private mapProgramDataElementToQuestions(
-        isTrackerProgram: boolean,
-        sectionDataElements: { id: string }[],
-        dataElements: ProgramDataElement[],
-        options: Option[],
-        event: D2TrackerEvent | undefined = undefined,
-        trackedEntity: TrackedEntity | undefined = undefined
-    ): { questions: Question[]; sectionAddQuestion: string } {
-        let sectionAddQuestion = "";
-        const questions: Question[] = _(
-            sectionDataElements.map(dataElement => {
-                const curDataElement = dataElements.find(de => de.id === dataElement.id);
-
-                if (curDataElement) {
-                    let dataValue: string | undefined;
-                    if (isTrackerProgram) {
-                        const dataValues = trackedEntity?.enrollments?.flatMap(enrollment => {
-                            return _(
-                                enrollment.events.map(e => {
-                                    return e.dataValues.find(
-                                        dv => dv.dataElement === curDataElement.id
-                                    );
-                                })
-                            )
-                                .compact()
-                                .value();
-                        });
-
-                        if (dataValues && dataValues.length > 1) {
-                            console.debug(
-                                "ERROR : There should never be more than one instance of a dataelement in an event"
-                            );
-                        }
-
-                        dataValue = dataValues?.at(0)?.value;
-                    } else {
-                        dataValue = event
-                            ? event.dataValues.find(dv => dv.dataElement === curDataElement.id)
-                                  ?.value
-                            : undefined;
-                    }
-
-                    const currentQuestion = this.getQuestion(
-                        curDataElement.valueType,
-                        curDataElement.id,
-                        curDataElement.code,
-                        curDataElement.formName,
-                        curDataElement.sortOrder,
-                        options,
-                        curDataElement.optionSet,
-                        dataValue ?? ""
-                    );
-
-                    ///Disable Id fields which are auto generated.
-                    if (
-                        currentQuestion &&
-                        (currentQuestion.id === SURVEY_ID_DATAELEMENT_ID ||
-                            currentQuestion.id === SURVEY_ID_PATIENT_DATAELEMENT_ID ||
-                            currentQuestion.id === WARD_ID_DATAELEMENT_ID ||
-                            currentQuestion.id === WARD2_ID_DATAELEMENT_ID ||
-                            currentQuestion.id === AMR_SURVEYS_PREVALENCE_DEA_SURVEY_ID)
-                    ) {
-                        currentQuestion.disabled = true;
-                    }
-
-                    //Some field was hidden, set it as label
-                    if (currentQuestion?.isVisible === false) {
-                        const LsectionAddLabel = hiddenFields.find(
-                            field => field === currentQuestion.text
-                        );
-
-                        if (LsectionAddLabel) {
-                            sectionAddQuestion = currentQuestion.id;
-                        }
-                    }
-
-                    return currentQuestion;
-                }
-            })
-        )
-            .compact()
-            .sortBy(q => q.sortOrder)
-            .value();
-
-        return { questions, sectionAddQuestion: sectionAddQuestion };
-    }
-
-    private mapTrackedAttributesToQuestions(
-        attributes: TrackedEntityAttibute[],
-        options: Option[],
-        trackedEntity: TrackedEntity | undefined
-    ): Question[] {
-        const questions: Question[] = _(
-            attributes.map(attribute => {
-                const attributeValue = trackedEntity?.attributes?.find(
-                    attr => attr.attribute === attribute.id
-                );
-
-                const currentQuestion = this.getQuestion(
-                    attribute.valueType,
-                    attribute.id,
-                    attribute.code,
-                    attribute.formName,
-                    attribute.sortOrder,
-                    options,
-                    attribute.optionSet,
-                    attributeValue?.value
-                );
-                if (
-                    currentQuestion &&
-                    (currentQuestion.id === SURVEY_ID_FACILITY_LEVEL_DATAELEMENT_ID ||
-                        currentQuestion?.id === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SSTF ||
-                        currentQuestion.id === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRL ||
-                        currentQuestion.id === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_PIS ||
-                        currentQuestion.id === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SRL ||
-                        currentQuestion.id === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRF)
-                ) {
-                    currentQuestion.disabled = true;
-                }
-                return currentQuestion;
-            })
-        )
-            .compact()
-            .sortBy(q => q.sortOrder)
-            .value();
-
-        return questions;
-    }
-
-    private getQuestion(
-        valueType: string,
-        id: Id,
-        code: string,
-        formName: string,
-        sortOrder: number | undefined,
-        options: Option[],
-        optionSet?: { id: string },
-        dataValue?: string
-    ): Question | undefined {
-        switch (valueType) {
-            case "BOOLEAN": {
-                const boolQ: BooleanQuestion = {
-                    id: id,
-                    code: code, //code
-                    text: formName, //formName
-                    type: "boolean",
-                    storeFalse: true,
-                    value: dataValue ? (dataValue === "true" ? true : false) : true,
-                    isVisible: hiddenFields.some(field => field === formName) ? false : true,
-                    sortOrder: sortOrder,
-                    errors: [],
-                };
-                return boolQ;
-            }
-            case "TRUE_ONLY": {
-                const boolQ: BooleanQuestion = {
-                    id: id,
-                    code: code, //code
-                    text: formName, //formName
-                    type: "boolean",
-                    storeFalse: false,
-                    value: dataValue ? (dataValue === "true" ? true : undefined) : undefined,
-                    isVisible: hiddenFields.some(field => field === formName) ? false : true,
-                    sortOrder: sortOrder,
-                    errors: [],
-                };
-                return boolQ;
-            }
-
-            case "NUMBER":
-            case "INTEGER": {
-                const intQ: NumberQuestion = {
-                    id: id,
-                    code: code, //code
-                    text: formName, //formName
-                    type: "number",
-                    numberType: "INTEGER",
-                    value: dataValue ? dataValue : "",
-                    isVisible: hiddenFields.some(field => field === formName) ? false : true,
-                    sortOrder: sortOrder,
-                    errors: [],
-                };
-                return intQ;
-            }
-
-            case "PHONE_NUMBER":
-            case "EMAIL":
-            case "TEXT": {
-                if (optionSet) {
-                    const selectOptions = options.filter(op => op.optionSet.id === optionSet?.id);
-
-                    const selectedOption = dataValue
-                        ? selectOptions.find(o => o.code === dataValue)
-                        : undefined;
-
-                    const selectQ: SelectQuestion = {
-                        id: id || "",
-                        code: code || "",
-                        text: formName || "",
-                        type: "select",
-                        options: selectOptions,
-                        value: selectedOption ? selectedOption : { name: "", id: "", code: "" },
-                        isVisible: hiddenFields.some(field => field === formName) ? false : true,
-                        sortOrder: sortOrder,
-                        errors: [],
-                    };
-                    return selectQ;
-                } else {
-                    const singleLineText: TextQuestion = {
-                        id: id,
-                        code: code,
-                        text: formName,
-                        type: "text",
-                        value: dataValue ? (dataValue as string) : "",
-                        multiline: false,
-                        isVisible: hiddenFields.some(field => field === formName) ? false : true,
-                        sortOrder: sortOrder,
-                        errors: [],
-                    };
-                    return singleLineText;
-                }
-            }
-
-            case "LONG_TEXT": {
-                const singleLineTextQ: TextQuestion = {
-                    id: id,
-                    code: code,
-                    text: formName,
-                    type: "text",
-                    value: dataValue ? (dataValue as string) : "",
-                    multiline: true,
-                    isVisible: hiddenFields.some(field => field === formName) ? false : true,
-                    sortOrder: sortOrder,
-                    errors: [],
-                };
-                return singleLineTextQ;
-            }
-
-            case "DATE": {
-                const dateQ: DateQuestion = {
-                    id: id,
-                    code: code,
-                    text: formName,
-                    type: "date",
-                    value: dataValue ? new Date(dataValue as string) : new Date(),
-                    isVisible: hiddenFields.some(field => field === formName) ? false : true,
-                    sortOrder: sortOrder,
-                    errors: [],
-                };
-                return dateQ;
-            }
-
-            case "DATETIME": {
-                const dateQ: DateTimeQuestion = {
-                    id: id,
-                    code: code,
-                    text: formName,
-                    type: "datetime",
-                    value: dataValue
-                        ? new Date(dataValue as string).toISOString()
-                        : new Date().toISOString(),
-                    isVisible: hiddenFields.some(field => field === formName) ? false : true,
-                    sortOrder: sortOrder,
-                    errors: [],
-                };
-                return dateQ;
-            }
-        }
-    }
-
     saveFormData(
         questionnaire: Questionnaire,
         action: ImportStrategy,
@@ -638,8 +155,8 @@ export class SurveyD2Repository implements SurveyRepository {
         programId: Id
     ): FutureData<void> {
         const $payload = isTrackerProgram(programId)
-            ? this.mapQuestionnaireToTrackedEntities(questionnaire, orgUnitId, programId, eventId)
-            : this.mapQuestionnaireToEvent(questionnaire, orgUnitId, programId, eventId);
+            ? mapQuestionnaireToTrackedEntities(questionnaire, orgUnitId, programId, eventId)
+            : mapQuestionnaireToEvent(questionnaire, orgUnitId, programId, this.api, eventId);
 
         return $payload.flatMap(payload => {
             return apiToFuture(
@@ -661,155 +178,6 @@ export class SurveyD2Repository implements SurveyRepository {
                 });
             });
         });
-    }
-
-    private mapQuestionsToDataValues(questions: Question[]): DataValue[] {
-        const dataValues = _(
-            questions.map(question => {
-                if (question) {
-                    if (question.type === "select" && question.value) {
-                        return {
-                            dataElement: question.id,
-                            value: question.value.code,
-                        };
-                    } else if (question.type === "date" && question.value) {
-                        return {
-                            dataElement: question.id,
-                            value: question.value.toISOString().split("T")?.at(0) || "",
-                        };
-                    } else if (question.type === "boolean" && question.storeFalse === false) {
-                        return {
-                            dataElement: question.id,
-                            value: question.value ? question.value : undefined,
-                        };
-                    } else {
-                        return {
-                            dataElement: question.id,
-                            value: question.value,
-                        };
-                    }
-                }
-            })
-        )
-            .compact()
-            .value();
-
-        return dataValues as DataValue[];
-    }
-
-    private mapQuestionnaireToEvent(
-        questionnaire: Questionnaire,
-        orgUnitId: string,
-        programId: Id,
-        eventId: string | undefined = undefined
-    ): FutureData<{ events: D2TrackerEvent[] }> {
-        const questions = questionnaire.stages.flatMap(stages =>
-            stages.sections.flatMap(section => section.questions)
-        );
-
-        const dataValues = this.mapQuestionsToDataValues(questions);
-
-        if (eventId) {
-            return this.getEventProgramById(eventId).flatMap(event => {
-                if (event) {
-                    const updatedEvent: D2TrackerEvent = {
-                        ...event,
-
-                        dataValues: dataValues as DataValue[],
-                    };
-                    return Future.success({ events: [updatedEvent] });
-                } else {
-                    return Future.error(new Error("Unable to find event to update"));
-                }
-            });
-        } else {
-            const event: D2TrackerEvent = {
-                event: "",
-                orgUnit: orgUnitId,
-                program: programId,
-                status: "ACTIVE",
-                occurredAt: new Date().toISOString().split("T")?.at(0) || "",
-                //@ts-ignore
-                dataValues: dataValues,
-            };
-            return Future.success({ events: [event] });
-        }
-    }
-
-    private mapQuestionnaireToTrackedEntities(
-        questionnaire: Questionnaire,
-        orgUnitId: string,
-        programId: Id,
-        teiId: string | undefined = undefined
-    ): FutureData<{ trackedEntities: TrackedEntity[] }> {
-        const eventsByStage: D2TrackerEvent[] = questionnaire.stages.map(stage => {
-            const dataValuesByStage = stage.sections.flatMap(section => {
-                return this.mapQuestionsToDataValues(section.questions);
-            });
-
-            return {
-                program: programId,
-                event: stage.instanceId ?? "",
-                programStage: stage.code,
-                orgUnit: orgUnitId,
-                dataValues: dataValuesByStage as DataValue[],
-                occurredAt: new Date().getTime().toString(),
-                status: "ACTIVE",
-            };
-        });
-
-        const attributes: D2TrackerEnrollmentAttribute[] = questionnaire.entity
-            ? questionnaire.entity.questions.map(question => {
-                  if (question.type === "select" && question.value) {
-                      return {
-                          attribute: question.id,
-                          value: question.value.code ? question.value.code : "",
-                      };
-                  } else if (question.type === "date") {
-                      return {
-                          attribute: question.id,
-                          value: question.value ? question.value : new Date(),
-                      };
-                  } else {
-                      return {
-                          attribute: question.id,
-                          value: question.value ? question.value.toString() : "",
-                      };
-                  }
-              })
-            : [];
-
-        const enrollments: D2TrackerEnrollment[] = [
-            {
-                orgUnit: orgUnitId,
-                program: programId,
-                enrollment: questionnaire.subLevelDetails?.enrollmentId ?? "",
-                trackedEntityType: getTrackedEntityAttributeType(programId),
-                notes: [],
-                relationships: [],
-                attributes: attributes,
-                events: eventsByStage,
-                enrolledAt: new Date().getTime().toString(),
-                occurredAt: new Date().getTime().toString(),
-                createdAt: new Date().getTime().toString(),
-                createdAtClient: new Date().getTime().toString(),
-                updatedAt: new Date().getTime().toString(),
-                updatedAtClient: new Date().getTime().toString(),
-                status: "ACTIVE",
-                orgUnitName: "",
-                followUp: false,
-                deleted: false,
-                storedBy: "",
-            },
-        ];
-
-        const entity: TrackedEntity = {
-            orgUnit: orgUnitId,
-            trackedEntity: teiId ?? "",
-            trackedEntityType: getTrackedEntityAttributeType(programId),
-            enrollments: enrollments,
-        };
-        return Future.success({ trackedEntities: [entity] });
     }
 
     getSurveys(
@@ -841,47 +209,8 @@ export class SurveyD2Repository implements SurveyRepository {
                 ouMode: ouMode,
             })
         ).flatMap((trackedEntities: TrackedEntitiesGetResponse) => {
-            const surveys = this.mapTrackedEntityToSurvey(trackedEntities, surveyFormType);
+            const surveys = mapTrackedEntityToSurvey(trackedEntities, surveyFormType);
             return Future.success(surveys);
-        });
-    }
-
-    mapTrackedEntityToSurvey(
-        trackedEntities: TrackedEntitiesGetResponse,
-        surveyFormType: SURVEY_FORM_TYPES
-    ) {
-        return trackedEntities.instances.map(trackedEntity => {
-            const parentPrevalenceSurveyId =
-                trackedEntity.attributes?.find(
-                    attribute =>
-                        attribute.attribute === SURVEY_ID_FACILITY_LEVEL_DATAELEMENT_ID ||
-                        attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SSTF ||
-                        attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRL ||
-                        attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_PIS ||
-                        attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_SRL ||
-                        attribute.attribute === AMR_SURVEYS_PREVALENCE_TEA_SURVEY_ID_CRF
-                )?.value ?? "";
-
-            const survey: Survey = {
-                id: trackedEntity.trackedEntity ?? "",
-                name: trackedEntity.trackedEntity ?? "",
-                rootSurvey: {
-                    id: parentPrevalenceSurveyId ?? "",
-                    name: "",
-                    surveyType: "",
-                },
-                startDate: trackedEntity.createdAt ? new Date(trackedEntity.createdAt) : undefined,
-                status: "ACTIVE",
-                assignedOrgUnit: {
-                    id: trackedEntity.orgUnit ?? "",
-                    name: "",
-                },
-                surveyType: "",
-                parentWardRegisterId: undefined,
-                surveyFormType: surveyFormType,
-                childCount: undefined,
-            };
-            return survey;
         });
     }
 
@@ -906,82 +235,8 @@ export class SurveyD2Repository implements SurveyRepository {
             })
         ).flatMap(response => {
             const events = response.instances;
-            const surveys = this.mapEventToSurvey(events, surveyFormType, programId);
+            const surveys = mapEventToSurvey(events, surveyFormType, programId);
             return Future.success(surveys);
-        });
-    }
-
-    mapEventToSurvey(
-        events: D2TrackerEvent[],
-        surveyFormType: SURVEY_FORM_TYPES,
-        programId: Id
-    ): Survey[] {
-        return events.map((event: D2TrackerEvent) => {
-            const surveyProperties = new Map(
-                keyToDataElementMap.map(({ key, dataElements }) => {
-                    const value =
-                        event.dataValues.find(dv => dataElements.includes(dv.dataElement))?.value ??
-                        "";
-
-                    return [key, value] as const;
-                })
-            );
-
-            const startDateStr = surveyProperties.get("startDate");
-            const startDate = startDateStr ? new Date(startDateStr) : undefined;
-            const surveyName = surveyProperties.get("surveyName") ?? "";
-            const surveyCompleted = surveyProperties.get("surveyCompleted") ?? "";
-            const hospitalCode = surveyProperties.get("hospitalCode") ?? "";
-            const wardCode = surveyProperties.get("wardCode") ?? "";
-            const patientCode = surveyProperties.get("patientCode") ?? "";
-            const parentPPSSurveyId = surveyProperties.get("parentPPSSurveyId") ?? "";
-            const surveyType = surveyProperties.get("surveyType") ?? "";
-            const parentWardRegisterId = surveyProperties.get("parentWardRegisterId") ?? "";
-
-            const status =
-                surveyCompleted === "false" && startDate
-                    ? startDate > new Date()
-                        ? "FUTURE"
-                        : "ACTIVE"
-                    : "COMPLETED";
-
-            const survey: Survey = {
-                id: event.event,
-                name: getSurveyNameBySurveyFormType(surveyFormType, {
-                    eventId: event.event,
-                    surveyName,
-                    orgUnitName: event.orgUnitName,
-                    hospitalCode,
-                    wardCode,
-                    patientCode,
-                }),
-                rootSurvey: {
-                    id:
-                        surveyFormType !== "PPSSurveyForm" &&
-                        surveyFormType !== "PrevalenceSurveyForm"
-                            ? parentPPSSurveyId
-                            : event.event,
-                    name:
-                        surveyFormType !== "PPSSurveyForm" &&
-                        surveyFormType !== "PrevalenceSurveyForm"
-                            ? ""
-                            : surveyName,
-                    surveyType: surveyFormType === "PPSSurveyForm" ? surveyType : "",
-                },
-                startDate: startDate,
-                status:
-                    programId === PPS_SURVEY_FORM_ID || programId === PREVALENCE_SURVEY_FORM_ID
-                        ? status
-                        : event.status === "COMPLETED"
-                        ? ("COMPLETED" as SURVEY_STATUSES)
-                        : ("ACTIVE" as SURVEY_STATUSES),
-                assignedOrgUnit: { id: event.orgUnit, name: event.orgUnitName ?? "" },
-                surveyType: surveyType,
-                parentWardRegisterId: parentWardRegisterId,
-                surveyFormType: surveyFormType,
-                childCount: undefined,
-            };
-            return survey;
         });
     }
 
@@ -1045,66 +300,6 @@ export class SurveyD2Repository implements SurveyRepository {
                 } else return Future.success("");
             })
             .flatMapError(_err => Future.success(""));
-    }
-    private getProgramRules(
-        programRulesResponse: D2ProgramRule[] | undefined,
-        programRuleVariables: D2ProgramRuleVariable[] | undefined,
-        programRuleActionsResponse: D2ProgramRuleAction[] | undefined
-    ): QuestionnaireRule[] {
-        return (
-            programRulesResponse?.map(({ id, condition, programRuleActions: actions }) => {
-                const dataElementIds =
-                    condition.match(/#{(.*?)}/g)?.map(programRuleVariableName => {
-                        const variableName = programRuleVariableName.replace(/#{|}/g, "");
-
-                        const dataElementId =
-                            programRuleVariables?.find(
-                                programRuleVariable => variableName === programRuleVariable.name
-                            )?.dataElement?.id || "";
-
-                        return dataElementId;
-                    }) || [];
-
-                const parsedCondition = condition.replace(/#{(.*?)}/g, (match, programRuleVar) => {
-                    const dataElementId =
-                        programRuleVariables?.find(
-                            programRuleVariable => programRuleVariable.name === programRuleVar
-                        )?.dataElement?.id || "";
-
-                    return `#{${dataElementId}}`;
-                });
-
-                const programRuleActionIds: string[] = actions.map(action => action.id);
-
-                const programRuleActions: D2ProgramRuleAction[] | undefined =
-                    programRuleActionsResponse
-                        ?.filter(programRuleAction =>
-                            programRuleActionIds.includes(programRuleAction.id)
-                        )
-                        .map(programRuleAction => {
-                            return {
-                                id: programRuleAction.id,
-                                programRuleActionType: programRuleAction.programRuleActionType,
-                                data: programRuleAction.data,
-                                dataElement: programRuleAction.dataElement,
-                                programStageSection: {
-                                    id: programRuleAction.programStageSection?.id,
-                                },
-                                programStage: {
-                                    id: programRuleAction.programStage?.id,
-                                },
-                                content: programRuleAction.content,
-                            };
-                        });
-
-                return {
-                    id: id,
-                    condition: parsedCondition,
-                    dataElementIds: _(dataElementIds).uniq().value(),
-                    actions: programRuleActions || [],
-                };
-            }) || []
-        );
     }
 
     getSurveyChildCount(
